Add files via upload
This project undertakes an examination of a census conducted in a small town situated between two larger cities connected by a motorway. The primary objective is to conduct a comprehensive analysis of the population and explore the most effective utilization of an unused land parcel in the area. Initially presented as raw data fraught with errors, the census data poses challenges for analysis, necessitating the initial step of refining it into more usable and accurate data. This cleansing process will be addressed in the initial section of the report